requirements.txt
blinker==1.6.2
certifi==2024.02.02
charset-normalizer==2.1.1
click==8.1.7
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
PyMySQL==1.0.2
python-dotenv==0.21.1
SQLAlchemy==1.4.46
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
requests==2.31.0
urllib3==1.26.18
Werkzeug==2.0.3
